requirements.txt
django~=5.1.1
pandas~=2.2.2
whitenoise
xlsxwriter
django-admin-logs
django_select2
requests~=2.32.3
xhtml2pdf~=0.2.16
pdfkit
social-auth-app-django
django-modeltranslation~=0.19.19
isort
ruff
sshtunnel~=0.4.0